Introduce docstrings and revise CI
In addition to docstrings changes, following enhancments were made:
* add `make-test` Makefile rule to perform test coverage. The same rule
  is executed in `main.yml`
* remove GetFX import from __init__
* move cmd logic from __main__ to getfxnbp module. This impacts changes
  / additions to tests to verify if this logic works correctly.
* omit `__main__.py` in testing coverage
* add in `.gitignore` `htmlconv/` folder and .DS_Store
* add tests to verify if there is correct network connection and if
  correct commandline attributes are provided
* remove obsolete test makred as `xfail` that incorrectly used mock
  (patch)

Additionally to above a pretty important fix was made that is related to
fixing issue of testing default currency:

For some reason ptw and pytest worked correctly with test:
`test_cmd_parser.py:test_empy_currency`, but when pytest was executed
with `-s` or `-v` switch it was not passing. In order to fix it instead
of `None`, emptry list was used in test `[]`.