Add Publish step in workflow with condition
It is mandatory to apply python version condition, because by default if
strategy matrix of python versions is used, documentation is pushed
in parallel to to single repository. This may randomly cause `cannot
lock` error during pushing documentation to remote - see example:
```
 ! [remote rejected] github-pages-deploy-action/yz4w8tkx4 -> gh-pages (cannot lock ref 'refs/heads/gh-pages': is at e2323442aae2926ff079b5ac9486983ff80d53c0 but expected eba0953dd875927a5fc16e519a3ca93087eba867)
```

Therefore it is important to perform check of Python version and perform
publishing only for pre-defined python version (e.g. 3.9)

*Note*: during work on this commit it was found that curly braces when
`if` expression evaluation should be used with care. Initially `if`
statement used `if: ${{ matrix.python-version }} == '3.9'` but this
appears not work correctly and `if` statement was alwys true. After `$`
and curly braces were removed the `if` statement worked as expected. It
appears that curly braces are not required, as they are needed only for
expression evaluation - see: [Abount context and
expressions](https://docs.github.com/en/actions/reference/context-and-expression-syntax-for-github-actions)